refactor: 보드 생성 로직 변경
