+++ src/main/java/chess/controller/ChessController.java
@@ -1,6 +1,7 @@
 package chess.controller;
 
 import chess.domain.board.Board;
+import chess.domain.board.maker.StartingPiecesGenerator;
 import chess.view.InputView;
 import chess.view.OutputView;
 
@@ -18,7 +19,7 @@ public class ChessController {
         outputView.printGameStartGuideMessage();
         final GameCommand command = GameCommand.findBy(inputView.readGameCommand());
         if (command == GameCommand.START) {
-            final Board board = Board.createBoardWithStartingPieces();
+            final Board board = Board.createBoardWith(new StartingPiecesGenerator());
             outputView.printBoard(board.getPieces());
         }
     }
