+++ src/main/java/chess/domain/board/maker/PiecesGenerator.java
@@ -1,0 +1,10 @@
+package chess.domain.board.maker;
+
+import chess.domain.piece.Piece;
+
+import java.util.List;
+
+public interface PiecesGenerator {
+
+    List<Piece> generate();
+}
