+++ src/test/java/chess/TestPiecesGenerator.java
@@ -1,0 +1,21 @@
+package chess;
+
+import chess.domain.board.maker.PiecesGenerator;
+import chess.domain.piece.Piece;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class TestPiecesGenerator implements PiecesGenerator {
+
+    private final List<Piece> pieces;
+
+    public TestPiecesGenerator(final List<Piece> pieces) {
+        this.pieces = new ArrayList<>(pieces);
+    }
+
+    @Override
+    public List<Piece> generate() {
+        return pieces;
+    }
+}
